Split backend agent model boundary
